refactor(resource): move utilization stats handling to booking domain and update related services
