+++ resource-app/backend/internal/booking/repository.go
@@ -12,12 +12,21 @@ type resourceLock struct {
 	ID string `gorm:"column:id"`
 }
 
+type utilizationRow struct {
+	ResourceID   string  `gorm:"column:resource_id"`
+	ResourceName string  `gorm:"column:resource_name"`
+	ResourceType string  `gorm:"column:resource_type"`
+	BookingCount int     `gorm:"column:booking_count"`
+	TotalHours   float64 `gorm:"column:total_hours"`
+}
+
 type Repository interface {
 	GetBookings() ([]Booking, error)
 	CreateBooking(booking *Booking) error
 	UpdateBookingStatus(id string, status BookingStatus, rejectionReason *string) (*Booking, error)
 	RescheduleBooking(id string, newStart, newEnd time.Time) (*Booking, error)
 	CancelBooking(id string) error
+	GetUtilizationStats() ([]ResourceUsageStats, error)
 }
 
 // GormRepository implements Repository using GORM
@@ -174,3 +183,46 @@ func (r *GormRepository) CancelBooking(id string) error {
 	}
 	return nil
 }
+
+func (r *GormRepository) GetUtilizationStats() ([]ResourceUsageStats, error) {
+	var rows []utilizationRow
+
+	err := r.db.Raw(`
+		SELECT
+			r.id   AS resource_id,
+			r.name AS resource_name,
+			r.type AS resource_type,
+			COUNT(b.id) AS booking_count,
+			COALESCE(SUM(TIMESTAMPDIFF(SECOND, b.start, b.end)), 0) / 3600 AS total_hours
+		FROM resources r
+		LEFT JOIN bookings b
+			ON b.resource_id = r.id AND b.status = ?
+		GROUP BY r.id, r.name, r.type
+	`, StatusConfirmed).Scan(&rows).Error
+	if err != nil {
+		return nil, err
+	}
+
+	stats := make([]ResourceUsageStats, 0, len(rows))
+	for _, row := range rows {
+		totalHours := int(row.TotalHours)
+		utilizationRate := 0
+		if totalHours > 0 {
+			utilizationRate = int((float64(totalHours) / 160.0) * 100.0)
+			if utilizationRate > 100 {
+				utilizationRate = 100
+			}
+		}
+		stats = append(stats, ResourceUsageStats{
+			ResourceID:      row.ResourceID,
+			ResourceName:    row.ResourceName,
+			ResourceType:    row.ResourceType,
+			BookingCount:    row.BookingCount,
+			TotalHours:      totalHours,
+			UtilizationRate: utilizationRate,
+		})
+	}
+
+	return stats, nil
+}
+
