+++ resource-app/backend/internal/resource/services.go
@@ -31,6 +31,3 @@ func (s *Service) GetResourceByID(id string) (*Resource, error) {
 	return s.repo.GetResourceByID(id)
 }
 
-func (s *Service) GetUtilizationStats() ([]ResourceUsageStats, error) {
-	return s.repo.GetUtilizationStats()
-}
